requirements.txt
Django==3.2.10
channels==3.0.4
daphne==3.0.2
djangorestframework==3.12.4
flake8==4.0.1
black ./core
